requirements.txt
paho-mqtt>=1.5.0
phatbeat>=0.1.1
pyalsaaudio>=0.9.0
RPi.GPIO>=0.7.0
dynaconf>=0.6.0
fastapi>=0.63.0
uvicorn>=0.13.4
pyinotify>=0.9.6
cysystemd>=1.5.2